SecureCheck
The Project focuses on Finding and Eliminating those threats on the basis of their md5hashes. It Collects items with over 70 antivirus scanners and URL/domain blacklisting services. Any user can select a file from their computer using their browser and input it to Secure-Check. It works on Python framework