Add files via upload
This simple Airhockey Game project is written in Python. The project file contains image files and python scripts. GUI uses Kivy module. A simple and clean GUI is provided for easy gameplay. The gameplay design is so simple that the user won’t find it difficult to use and understand. It has touch effects so you can control paddles easily.